feat: add function delete car
