+++ src/index.ts
@@ -73,8 +73,4 @@ rsImg.src = 'https://rs.school/images/rs_school_js.svg';
 rsImg.alt = 'Rolling Scopes School';
 rsLink.append(rsImg);
 
-(async () => {
-  console.log(await getCountCars())
-})();
-
 createPageGarage();
